Fix some things after rebasing on master
